Rename nbserverproxy to jupyter-server-proxy
