+++ jupyter_server_proxy/etc/nbextension.json
@@ -1,0 +1,6 @@
+{
+  "load_extensions": {
+     "jupyter_server_proxy/tree": true
+  }
+}
+
