+++ jupyter_server_proxy/etc/serverextension.json
@@ -1,7 +1,7 @@
 {
   "NotebookApp": {
     "nbserver_extensions": {
-      "nbserverproxy": true
+      "jupyter_server_proxy": true
     }
   }
 }
