+++ jupyterlab-server-proxy/src/index.ts
@@ -26,10 +26,10 @@ function addLauncherEntries(serverData: any, launcher: ILauncher, app: JupyterLa
     }
 }
 /**
- * Initialization data for the jupyterlab-nbserverproxy extension.
+ * Initialization data for the jupyterlab-server-proxy extension.
  */
 const extension: JupyterLabPlugin<void> = {
-  id: 'jupyterlab-nbserverproxy',
+  id: 'jupyterlab-server-proxy',
   autoStart: true,
   requires: [ILauncher],
   activate: (app: JupyterLab, launcher: ILauncher) => {
@@ -37,7 +37,7 @@ const extension: JupyterLabPlugin<void> = {
     fetch(PageConfig.getBaseUrl() + 'server-proxy/servers-info').then(
       response => {
         if (!response.ok) {
-          console.log('Fetching metadata about registered failed. Make sure nbserverproxy is installed');
+          console.log('Fetching metadata about registered failed. Make sure jupyter-server-proxy is installed');
           console.log(response);
         } else {
           response.json().then(data => addLauncherEntries(data, launcher, app))
