--- nbserverproxy/etc/nbextension.json
@@ -1,6 +1,0 @@
-{
-  "load_extensions": {
-     "nbserverproxy/tree": true
-  }
-}
-
